+++ tests/loops/optimization/test_optimizer_loop.py
@@ -11,16 +11,23 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+
+from torch.optim import Adam, SGD
+
+from pytorch_lightning.core.optimizer import LightningOptimizer
+from pytorch_lightning.loops.optimization.optimizer_loop import ClosureResult
+
+import os
+from unittest import mock
 from unittest.mock import Mock
 
 import pytest
 import torch
-from torch.optim import Adam, SGD
 
-from pytorch_lightning import Trainer
-from pytorch_lightning.core.optimizer import LightningOptimizer
-from pytorch_lightning.loops.optimization.optimizer_loop import ClosureResult
+from pytorch_lightning import seed_everything, Trainer
 from tests.helpers import BoringModel
+from tests.helpers.runif import RunIf
 
 
 def test_closure_result_deepcopy():
@@ -109,3 +116,102 @@ def test_optimizer_frequencies(tmpdir, frequencies, expected):
     assert all(isinstance(opt, LightningOptimizer) for opt in pl_optimizer_sequence)
     optimizer_sequence = [opt._optimizer.__class__.__name__ for opt in pl_optimizer_sequence]
     assert list(zip(opt_idx_sequence, optimizer_sequence)) == expected
+
+
+@RunIf(min_torch="1.7.0")
+@mock.patch.dict(os.environ, {"PL_FAULT_TOLERANT_TRAINING": "1"})
+@pytest.mark.parametrize("stop_epoch", (0, 1))
+@pytest.mark.parametrize("stop_batch", (0, 1, 2))
+@pytest.mark.parametrize("n_optimizers,stop_optimizer", [(2, 0), (2, 1), (3, 2)])
+def test_loop_restart_progress_multiple_optimizers(tmpdir, n_optimizers, stop_optimizer, stop_epoch, stop_batch):
+    n_batches = 3
+    n_epochs = 2
+    fail = False
+
+    def _assert_optimizer_sequence(method_mock, expected):
+        positional_args = [c[0] for c in method_mock.call_args_list]
+        sequence = [arg[3] for arg in positional_args]
+        assert sequence == expected
+
+    num_optimizers_incomplete = stop_epoch * n_batches * n_optimizers + stop_batch * n_optimizers + stop_optimizer
+
+    opt_idx_sequence_complete = list(range(n_optimizers)) * n_epochs * n_batches  # [0, 1, 2, 0, 1, 2, 0, 1, ...]
+    # +1 because we fail inside the closure inside optimizer_step()
+    opt_idx_sequence_incomplete = opt_idx_sequence_complete[: (num_optimizers_incomplete + 1)]
+    opt_idx_sequence_resumed = opt_idx_sequence_complete[num_optimizers_incomplete:]
+
+    class MultipleOptimizerModel(BoringModel):
+        def __init__(self):
+            super().__init__()
+
+        def training_step(self, batch, batch_idx, optimizer_idx):
+            if (
+                fail
+                and self.current_epoch == stop_epoch
+                and batch_idx == stop_batch
+                and optimizer_idx == stop_optimizer
+            ):
+                raise CustomException
+            return super().training_step(batch, batch_idx)
+
+        def configure_optimizers(self):
+            return [torch.optim.SGD(self.parameters(), lr=0.1) for _ in range(n_optimizers)]
+
+    # run without a failure, collect weights
+    fail = False
+    seed_everything(0)
+    model = MultipleOptimizerModel()
+    model.training_epoch_end = None
+    model.optimizer_step = Mock(wraps=model.optimizer_step)
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_epochs=n_epochs,
+        limit_train_batches=n_batches,
+        num_sanity_val_steps=0,
+        logger=False,
+    )
+    trainer.fit(model)
+    weights_complete = model.parameters()
+    _assert_optimizer_sequence(model.optimizer_step, opt_idx_sequence_complete)
+
+    # simulate a failure
+    fail = True
+    seed_everything(0)
+    model = MultipleOptimizerModel()
+    model.training_epoch_end = None
+    model.optimizer_step = Mock(wraps=model.optimizer_step)
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_epochs=n_epochs,
+        limit_train_batches=n_batches,
+        num_sanity_val_steps=0,
+        logger=False,
+    )
+    with pytest.raises(CustomException):
+        trainer.fit(model)
+
+    _assert_optimizer_sequence(model.optimizer_step, opt_idx_sequence_incomplete)
+
+    # resume from failure and collect weights
+    fail = False
+    seed_everything(0)
+    model = MultipleOptimizerModel()
+    model.training_epoch_end = None
+    model.optimizer_step = Mock(wraps=model.optimizer_step)
+    trainer = Trainer(
+        resume_from_checkpoint=str(tmpdir / ".pl_auto_save.ckpt"),
+        default_root_dir=tmpdir,
+        max_epochs=n_epochs,
+        limit_train_batches=n_batches,
+        limit_val_batches=0,
+        num_sanity_val_steps=0,
+        logger=False,
+    )
+    trainer.fit(model)
+    weights_resumed = model.parameters()
+
+    # check that the final weights of a resumed run match the weights of a run that never failed
+    for w0, w1 in zip(weights_complete, weights_resumed):
+        assert torch.allclose(w0, w1)
+
+    _assert_optimizer_sequence(model.optimizer_step, opt_idx_sequence_resumed)
