+++ tests/loops/test_loops.py
@@ -16,12 +16,12 @@ from copy import deepcopy
 from dataclasses import dataclass
 from typing import Any, Dict, Iterator
 from unittest import mock
-from unittest.mock import ANY, Mock
+from unittest.mock import ANY
 
 import pytest
 import torch
 
-from pytorch_lightning import seed_everything, Trainer
+from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import ModelCheckpoint
 from pytorch_lightning.loops import Loop, TrainingBatchLoop
 from pytorch_lightning.trainer.progress import BaseProgress
@@ -330,105 +330,6 @@ def test_loop_restart_progress_multiple_dataloaders(tmpdir, n_dataloaders, stop_
         "current": {"ready": stop_batch, "started": stop_batch, "processed": stop_batch, "completed": stop_batch},
     }
     assert trainer.fit_loop.epoch_loop.val_loop.epoch_loop.batch_progress.state_dict() == expected
-
-
-@RunIf(min_torch="1.7.0")
-@mock.patch.dict(os.environ, {"PL_FAULT_TOLERANT_TRAINING": "1"})
-@pytest.mark.parametrize("stop_epoch", (0, 1))
-@pytest.mark.parametrize("stop_batch", (0, 1, 2))
-@pytest.mark.parametrize("n_optimizers,stop_optimizer", [(2, 0), (2, 1), (3, 2)])
-def test_loop_restart_progress_multiple_optimizers(tmpdir, n_optimizers, stop_optimizer, stop_epoch, stop_batch):
-    n_batches = 3
-    n_epochs = 2
-    fail = False
-
-    def _assert_optimizer_sequence(method_mock, expected):
-        positional_args = [c[0] for c in method_mock.call_args_list]
-        sequence = [arg[3] for arg in positional_args]
-        assert sequence == expected
-
-    num_optimizers_incomplete = stop_epoch * n_batches * n_optimizers + stop_batch * n_optimizers + stop_optimizer
-
-    opt_idx_sequence_complete = list(range(n_optimizers)) * n_epochs * n_batches  # [0, 1, 2, 0, 1, 2, 0, 1, ...]
-    # +1 because we fail inside the closure inside optimizer_step()
-    opt_idx_sequence_incomplete = opt_idx_sequence_complete[: (num_optimizers_incomplete + 1)]
-    opt_idx_sequence_resumed = opt_idx_sequence_complete[num_optimizers_incomplete:]
-
-    class MultipleOptimizerModel(BoringModel):
-        def __init__(self):
-            super().__init__()
-
-        def training_step(self, batch, batch_idx, optimizer_idx):
-            if (
-                fail
-                and self.current_epoch == stop_epoch
-                and batch_idx == stop_batch
-                and optimizer_idx == stop_optimizer
-            ):
-                raise CustomException
-            return super().training_step(batch, batch_idx)
-
-        def configure_optimizers(self):
-            return [torch.optim.SGD(self.parameters(), lr=0.1) for _ in range(n_optimizers)]
-
-    # run without a failure, collect weights
-    fail = False
-    seed_everything(0)
-    model = MultipleOptimizerModel()
-    model.training_epoch_end = None
-    model.optimizer_step = Mock(wraps=model.optimizer_step)
-    trainer = Trainer(
-        default_root_dir=tmpdir,
-        max_epochs=n_epochs,
-        limit_train_batches=n_batches,
-        num_sanity_val_steps=0,
-        logger=False,
-    )
-    trainer.fit(model)
-    weights_complete = model.parameters()
-    _assert_optimizer_sequence(model.optimizer_step, opt_idx_sequence_complete)
-
-    # simulate a failure
-    fail = True
-    seed_everything(0)
-    model = MultipleOptimizerModel()
-    model.training_epoch_end = None
-    model.optimizer_step = Mock(wraps=model.optimizer_step)
-    trainer = Trainer(
-        default_root_dir=tmpdir,
-        max_epochs=n_epochs,
-        limit_train_batches=n_batches,
-        num_sanity_val_steps=0,
-        logger=False,
-    )
-    with pytest.raises(CustomException):
-        trainer.fit(model)
-
-    _assert_optimizer_sequence(model.optimizer_step, opt_idx_sequence_incomplete)
-
-    # resume from failure and collect weights
-    fail = False
-    seed_everything(0)
-    model = MultipleOptimizerModel()
-    model.training_epoch_end = None
-    model.optimizer_step = Mock(wraps=model.optimizer_step)
-    trainer = Trainer(
-        resume_from_checkpoint=str(tmpdir / ".pl_auto_save.ckpt"),
-        default_root_dir=tmpdir,
-        max_epochs=n_epochs,
-        limit_train_batches=n_batches,
-        limit_val_batches=0,
-        num_sanity_val_steps=0,
-        logger=False,
-    )
-    trainer.fit(model)
-    weights_resumed = model.parameters()
-
-    # check that the final weights of a resumed run match the weights of a run that never failed
-    for w0, w1 in zip(weights_complete, weights_resumed):
-        assert torch.allclose(w0, w1)
-
-    _assert_optimizer_sequence(model.optimizer_step, opt_idx_sequence_resumed)
 
 
 @RunIf(min_torch="1.7.0")
